+++ src/app/manifestGeneration/ManifestGeneratorFactory.ts
@@ -4,6 +4,7 @@
 
 import { ManifestVersions } from "./ManifestVersions";
 import { ManifestGenerator as ManifestGenerator13 } from "./manifestGenerators/generator13/ManifestGenerator";
+import { ManifestGenerator as ManifestGenerator14 } from "./manifestGenerators/generator14/ManifestGenerator";
 import { ManifestGenerator as ManifestGeneratorDevPreview } from "./manifestGenerators/generatorDevPreview/ManifestGenerator";
 import { BaseManifestGenerator } from "./BaseManifestGenerator";
 
@@ -13,12 +14,20 @@ export class ManifestGeneratorFactory {
         {
             manifestVersion: ManifestVersions.v13,
             schemaUrl: "https://developer.microsoft.com/en-us/json-schemas/teams/v1.3/MicrosoftTeams.schema.json",
-            manifestValue: "1.3"
+            manifestValue: "1.3",
+            default: false
+        },
+        {
+            manifestVersion: ManifestVersions.v14,
+            schemaUrl: "https://developer.microsoft.com/en-us/json-schemas/teams/v1.4/MicrosoftTeams.schema.json",
+            manifestValue: "1.4",
+            default: true
         },
         {
             manifestVersion: ManifestVersions.devPreview,
             schemaUrl: "https://raw.githubusercontent.com/OfficeDev/microsoft-teams-app-schema/preview/DevPreview/MicrosoftTeams.schema.json",
-            manifestValue: "devPreview"
+            manifestValue: "devPreview",
+            default: false
         }
     ]
 
@@ -40,6 +49,8 @@ export class ManifestGeneratorFactory {
     public createManifestGenerator(manifestVersion: string): BaseManifestGenerator {
         if (manifestVersion == ManifestVersions.v13) {
             return new ManifestGenerator13();
+        } else if (manifestVersion == ManifestVersions.v14) {
+            return new ManifestGenerator14();
         } else if (manifestVersion == ManifestVersions.devPreview) {
             return new ManifestGeneratorDevPreview();
         } else {