+++ src/app/templates/gulpfile.js
@@ -37,6 +37,10 @@ const SCHEMAS = [
         schema: "https://developer.microsoft.com/en-us/json-schemas/teams/v1.3/MicrosoftTeams.schema.json"
     },
     {
+        version: "1.4",
+        schema: "https://developer.microsoft.com/en-us/json-schemas/teams/v1.4/MicrosoftTeams.schema.json"
+    },
+    {
         version: "devPreview",
         schema: "https://raw.githubusercontent.com/OfficeDev/microsoft-teams-app-schema/preview/DevPreview/MicrosoftTeams.schema.json"
     }